Summarizing rating objects

Summarizing rating objects Summary for object of class rating

# S3 method for class 'rating'
summary(object, ...)

Arguments

object

of class rating

...

optional arguments

Value

List with following elements:

formula modeled formula.

method type of algorithm used.

Overall Accuracy named vector containing players ratings.

r a data.frame with summarized players ratings and model winning probabilities.

Probabilities are returned only in models with one variable (ratings):

name of a player

r players ratings

rd players ratings deviation

Model probability mean predicted probability of winning the challenge by the player.

True probability mean observed probability of winning the challenge by the player.

Accuracy Accuracy of prediction.

pairings number of pairwise occurrences.

Examples

model <- glicko_run(
  formula = rank | id ~ player(rider),
  data = gpheats[1:102, ]
)
summary(model)
#> $formula
#> rank | id ~ player(rider)
#> <environment: 0x55c4824ec048>
#> 
#> $method
#> [1] "glicko"
#> 
#> $`Overall Accuracy`
#> [1] 0.6066667
#> 
#> $`Number of pairs`
#> [1] 300
#> 
#> $r
#> Key: <rider>
#>                 rider        r      rd Model probability True probability
#>                <char>    <num>   <num>             <num>            <num>
#>  1:        Andy Smith 1380.985 106.955             0.422            0.389
#>  2:      Billy Hamill 1615.410 135.708             0.520            0.636
#>  3:       Chris Louis 1698.935 102.913             0.669            0.700
#>  4:       Craig Boyce 1518.764  99.693             0.457            0.500
#>  5:     Dariusz Śledź 1459.486 110.219             0.407            0.467
#>  6:     Gary Havelock 1436.409  95.033             0.468            0.429
#>  7:      Greg Hancock 1431.845 104.982             0.594            0.444
#>  8:      Hans Nielsen 1940.670 112.337             0.691            0.882
#>  9: Henrik Gustafsson 1470.559  99.201             0.501            0.444
#> 10:    Jan Staechmann 1287.756 111.101             0.282            0.278
#> 11:        Mark Loram 1621.466 103.416             0.692            0.619
#> 12:        Marvyn Cox 1281.862 100.506             0.323            0.250
#> 13:   Mikael Karlsson 1108.342 112.249             0.247            0.143
#> 14:    Peter Karlsson 1442.468 199.270             0.426            0.333
#> 15:     Sam Ermolenko 1698.492 105.125             0.642            0.667
#> 16:     Tomasz Gollob 1847.471 106.610             0.550            0.765
#> 17:     Tommy Knudsen 1109.247 173.454             0.595            0.167
#> 18:  Tony Rickardsson 1690.942 103.188             0.545            0.667
#>     Accuracy pairings
#>        <num>    <int>
#>  1:    0.556       18
#>  2:    0.455       11
#>  3:    0.550       20
#>  4:    0.450       20
#>  5:    0.667       15
#>  6:    0.619       21
#>  7:    0.333       18
#>  8:    0.706       17
#>  9:    0.722       18
#> 10:    0.778       18
#> 11:    0.571       21
#> 12:    0.750       20
#> 13:    0.810       21
#> 14:    0.333        3
#> 15:    0.500       18
#> 16:    0.706       17
#> 17:    0.333        6
#> 18:    0.611       18
#>