Summarizing rating objects

Summarizing rating objects Summary for object of class `rating`

# S3 method for rating
summary(object, ...)

Arguments

object
of class rating

...
optional arguments

Value

List with following elements

formula modeled formula.

method type of algorithm used.

Overall Accuracy named vector containing players ratings.

r data.frame summarized players ratings and model winning probabilities. Probabilities are returned only in models with one variable (ratings)

name of a player

r players ratings

rd players ratings deviation

`Model probability` mean predicted probability of winning the challenge by the player.

`True probability` mean observed probability of winning the challenge by the player.

`Accuracy` Accuracy of prediction.

`pairings` number of pairwise occurrences.

Examples

model <- glicko_run(formula = rank | id ~ player(rider),
                    data = gpheats[1:102, ])
summary(model)

#> $formula
#> rank | id ~ player(rider)
#> <environment: 0x5c441b0>
#> 
#> $method
#> [1] "glicko"
#> 
#> $`Overall Accuracy`
#> [1] 0.6133333
#> 
#> $`Number of pairs`
#> [1] 300
#> 
#> $r
#>                 rider        r      rd Model probability True probability
#>  1:        Andy SMITH 1380.985 106.955             0.422            0.389
#>  2:      Billy HAMILL 1615.410 135.708             0.520            0.636
#>  3:       Chris LOUIS 1725.243 101.653             0.646            0.700
#>  4:       Craig BOYCE 1457.914 105.573             0.459            0.444
#>  5:     Dariusz ŚLEDŹ 1459.486 110.219             0.407            0.467
#>  6:     Gary HAVELOCK 1436.409  95.033             0.468            0.429
#>  7:      Greg HANCOCK 1431.845 104.982             0.594            0.444
#>  8:      Hans NIELSEN 1940.670 112.337             0.691            0.882
#>  9: Henrik GUSTAFSSON 1470.559  99.201             0.501            0.444
#> 10:    Jan STAECHMANN 1287.756 111.101             0.282            0.278
#> 11:        Mark LORAM 1621.466 103.416             0.692            0.619
#> 12:        Marvyn COX 1295.448 103.035             0.310            0.250
#> 13:   Mikael KARLSSON 1108.342 112.249             0.247            0.143
#> 14:    Peter KARLSSON 1442.468 199.270             0.426            0.333
#> 15:     Sam ERMOLENKO 1698.492 105.125             0.642            0.667
#> 16:     Tomasz GOLLOB 1871.251 101.827             0.575            0.789
#> 17:     Tommy KNUDSEN 1109.247 173.454             0.595            0.167
#> 18:  Tony RICKARDSSON 1690.942 103.188             0.545            0.667
#>     Accuracy pairings
#>  1:    0.556       18
#>  2:    0.455       11
#>  3:    0.600       20
#>  4:    0.444       18
#>  5:    0.667       15
#>  6:    0.619       21
#>  7:    0.333       18
#>  8:    0.706       17
#>  9:    0.722       18
#> 10:    0.778       18
#> 11:    0.571       21
#> 12:    0.750       20
#> 13:    0.810       21
#> 14:    0.333        3
#> 15:    0.500       18
#> 16:    0.737       19
#> 17:    0.333        6
#> 18:    0.611       18
#>